pyproject.toml
[project]
name = "kube-chaos-lab"
version = "0.1.0"
description = "Reproducible local Kubernetes troubleshooting laboratory"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.0",
    "pyyaml"
    
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0"
]

[project.scripts]
labctl = "tools.labctl.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["tools*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
